# lista3/order/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/order-history/{id}/status":{"put":{"tags":["order-history-controller"],"operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["CREATED","PICKED_UP","DELIVERED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderHistory"}}}}}}},"/order-history":{"get":{"tags":["order-history-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderHistory"}}}}}}},"post":{"tags":["order-history-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderHistory"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderHistory"}}}}}}},"/order-history/sync":{"post":{"tags":["order-history-controller"],"operationId":"syncOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderHistory"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderHistory"}}}}}}},"/order-history/{id}":{"get":{"tags":["order-history-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderHistory"}}}}}}}},"components":{"schemas":{"OrderHistory":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customerName":{"type":"string"},"deliveryAddress":{"type":"string"},"products":{"type":"string"},"totalValue":{"type":"number","format":"double"},"status":{"type":"string","enum":["CREATED","PICKED_UP","DELIVERED"]}}}}}}